microservices/lstm_api/requirements.txt
fastapi==0.108.0
numpy==1.23.5
pandas==2.0.3
python-dateutil==2.8.2
TimeConvert==3.0.13
uvicorn==0.23.2
keras==2.15.0
packaging==23.2
Pillow==10.1.0
requests==2.31.0
scikit-learn==1.3.2
scipy==1.11.4
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
python-dotenv==1.0.0
matplotlib==3.8.2